+++ console/app.rb
@@ -18,7 +18,7 @@ class App
     @genres = [Genre.new('Comedy'), Genre.new('Thriller')]
     @labels = []
     @music_albums = load_music_album(@authors, @labels, @genres)
-    @games = load_games(@authors)
+    @games = load_games(@authors, @labels, @genres)
   end
 
   def list_all_books
