+++ data/games.json
@@ -1,1 +1,1 @@
-[]
+[{"publish_date":"a","archived":"z\n","multiplayer":null,"last_played_at":null,"author":"Stephen","genre":"Comedy","label":null},{"publish_date":"q","archived":"s\n","multiplayer":null,"last_played_at":null,"author":"Stephen","genre":"Comedy","label":null}]
